Fix for GitHub Pages no longer rending CSS correctly.
Renamed /assets/css/style.css to /assets/css/style.scss

Inserted “@import "{{ site.theme }}”;” on the third line.

Source:
https://help.github.com/articles/customizing-css-and-html-in-your-jekyll
-theme/

Likely cause of when the change was made:
https://github.com/blog/2289-publishing-with-github-pages-now-as-easy-as
-1-2-3

Discovered issue on July 6th 2017 when trying to post a new article on
my own fork of this project. Searched the projects issues and found #4
with the same problem. Found the fix shortly after in above articles.